HomePage with `About` and `GoToTest` buttons
Remove header and move the `AboutPage` button to HomePage
Load initial test data when the `GoToTest` button is clicked
