+++ src/App.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { render, fireEvent } from '@testing-library/react';
+import { render } from '@testing-library/react';
 
 import { MemoryRouter } from 'react-router-dom';
 
@@ -9,17 +9,6 @@ import { useDispatch, useSelector } from 'react-redux';
 import App from './App';
 
 import question from '../fixtures/question';
-
-jest.mock('react-redux');
-
-const mockPush = jest.fn();
-
-jest.mock('react-router-dom', () => ({
-  ...jest.requireActual('react-router-dom'),
-  useHistory() {
-    return { push: mockPush };
-  },
-}));
 
 describe('App', () => {
   const dispatch = jest.fn();
@@ -49,6 +38,7 @@ describe('App', () => {
       const { container } = renderApp({ path: '/' });
 
       expect(container).toHaveTextContent(/테스트하기/);
+      expect(container).toHaveTextContent(/About/);
     });
   });
 
@@ -65,16 +55,6 @@ describe('App', () => {
       const { container } = renderApp({ path: '/about' });
 
       expect(container).toHaveTextContent(/About/);
-    });
-  });
-
-  context('when click `임시제목`', () => {
-    it('occurs handleClickLink', () => {
-      const { getByText } = renderApp({ path: '/' });
-
-      fireEvent.click(getByText(/임시제목/));
-
-      expect(mockPush).toBeCalledWith('/');
     });
   });
 });
