+++ src/pages/HomePage.jsx
@@ -1,11 +1,22 @@
 import React from 'react';
 
-import { Link } from 'react-router-dom';
+import { Link, useHistory } from 'react-router-dom';
+
+import HomeContainer from '../containers/HomeContainer';
 
 export default function HomePage() {
+  const history = useHistory();
+
+  function handleClickLink(url) {
+    history.push(url);
+  }
+
   return (
     <div>
-      <Link to="/tests">테스트하기</Link>
+      <HomeContainer
+        handleClickLink={handleClickLink}
+      />
+      <Link to="/about">About</Link>
     </div>
   );
 }
